telomeres/src/tenx_telomeres/requirements.txt
pytest
discover
biopython
